refactor(config): split adaptive runtime model
